refactor(v0.69.0): tool_handlers.py 1472L → 15-module 패키지로 분할 (Tier 3 #1) (#892)
Tier 3 God Object 분할 시리즈 첫 번째. 1472 LOC monolith의 14개
_build_*_handlers() 팩토리를 각자 파일로 분리.

새 패키지 구조 — core/cli/tool_handlers/:
  __init__.py      148L  공개 _build_tool_handlers + _PLAN_STORE 싱글턴 + 19개 re-export
  _helpers.py       57L  _clarify, _safe_delegate, install_domain_tool_handlers
  memory.py         74L  memory_search/save, manage_rule
  plan.py          296L  create/approve/reject/modify/list_plan (최대 모듈)
  hitl.py          120L  rate/accept/reject_result, rerun_node
  system.py        250L  status/help/model/key/auth/login/doctor_slack
  execution.py     159L  schedule_job, trigger_event
  delegated.py      54L  _DELEGATED_TOOLS 레지스트리 + lazy 위임
  mcp.py            51L  install_mcp_server
  context.py        78L  manage_context (status/compact/clear)
  task.py          149L  task_create/update/get/list/stop
  notification.py   19L  send_notification
  calendar.py       33L  calendar_list/create_event/sync_scheduler
  offload.py        25L  recall_tool_result
  computer_use.py   27L  computer (조건부)

공개 API 변경 0건. 외부 import 7곳 (core/ 4 + tests/ 3) 수정 불필요 —
패키지 __init__.py가 19개 심볼 모두 re-export. test_plan_mode.py의
monkeypatch.setattr(th, "_PLAN_STORE", ...) 픽스처는 _PLAN_STORE를
패키지 루트에 유지해 그대로 동작.

Quality Gates:
- ruff/ruff-format/mypy 0 errors (260 source files, +14)
- pytest 4344 passed, 1 skipped, 24 deselected — 변동 없음
- E2E Cowboy Bebop A (68.4) 변동 없음
- 최대 단일 파일 1472 → 296 LOC (-80%)

남은 Tier 3 후보 8개 — commands.py(2441), cli/__init__.py(1889),
agent/loop.py(1754), scheduler.py(1208), agentic_ui.py(1160),
reports.py(1156), tool_executor.py(1047), llm/router.py(1046).
각각 별도 PR.
